+++ packages/share/src/global.d.ts
@@ -1,13 +1,9 @@
 /// <reference types="@types/alpinejs" />
 
-export type { ShareMagic, ShareStore } from "./index.js";
+export type { ShareMagic } from "./index.js";
 
 declare global {
   namespace Alpine {
-    interface Stores {
-      share: import("./index.js").ShareStore;
-    }
-
     interface Magics<T> {
       $share: import("./index.js").ShareMagic;
     }